determine the correct DAC divider value for generated filters
Previously it was assigning only static values for Tx instead of basing
the Tx DAC divider value off the ADC rate. Now we still assign static
values in cook_input (since it's currently a standalone design so the Tx
design doesn't know about the Rx values) but the values are sanity
checked and altered along with the PLL multiplier if they need to be
changed.

Signed-off-by: Tim Harder <timothy.harder@analog.com>